# drone-backend/src/main/java/com/uav/task/controller/UserTaskController.java
package com.uav.task.controller;

import com.fasterxml.jackson.core.type.TypeReference;
import com.uav.billing.service.BillConfigService;
import com.uav.server.annotation.RequireRole;
import com.uav.server.enums.Role;
import com.uav.task.pojo.dto.PriceEstimateDto;
import com.uav.task.pojo.dto.TaskDto;
import com.uav.task.pojo.entity.Task;
import com.uav.server.result.Result;
import com.uav.order.mapper.OrderRepository;
import com.uav.order.pojo.entity.MissionOrder;
import com.uav.task.mapper.TaskAssignmentRepository;
import com.uav.task.pojo.entity.TaskAssignment;
import com.uav.task.pojo.vo.AmapConfigVO;
import com.uav.task.pojo.vo.PriceDetailVO;
import com.uav.task.pojo.vo.PublishConfigVO;
import com.uav.task.pojo.vo.TaskActionHints;
import com.uav.task.pojo.vo.TaskPageVO;
import com.uav.task.pojo.vo.TaskVo;
import com.uav.server.annotation.OperationLog;
import com.uav.server.annotation.RateLimiter;
import com.uav.server.config.AmapConfig;
import com.uav.server.util.UserContext;
import com.uav.task.service.TaskService;
import com.uav.user.mapper.UserRepository;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.web.bind.annotation.*;

import java.math.BigDecimal;
import java.util.List;


@RequireRole({Role.USER, Role.ADMIN})
@Tag(name = "Task API", description = "任务创建与管理接口")
@RestController
@RequestMapping("/task")
@Slf4j
public class UserTaskController {

    @Autowired
    private AmapConfig amapConfig;

    @Autowired
    private TaskService taskService;

    @Autowired
    private BillConfigService billConfigService;

    @Autowired
    private TaskAssignmentRepository taskAssignmentRepository;

    @Autowired
    private OrderRepository orderRepository;

    @Autowired
    private UserRepository userRepository;

    @Autowired
    private com.uav.live.service.impl.LiveDeviceResolver liveDeviceResolver;

    @Operation(summary = "获取地图配置", description = "返回高德地图 JS API 所需的 key 与安全密钥")
    @GetMapping("/init")
    public Result<AmapConfigVO> init() {
        return Result.success(new AmapConfigVO(amapConfig.getKey(), amapConfig.getSecurityKey()));
    }

    @OperationLog("创建任务")
    @RateLimiter(limit = 10, windowSeconds = 60)
    @Operation(summary = "创建任务", description = "创建新任务，包含任务类型、航点信息等")
    @PostMapping("/create")
    public Result<TaskVo> createTask(@RequestBody TaskDto dto) {
        Task saved = taskService.createTask(dto);
        return Result.success("任务创建成功", toTaskVo(saved));
    }

    @OperationLog("参考价预估")
    @RateLimiter(limit = 30, windowSeconds = 60)
    @Operation(summary = "参考价预估", description = "发布前预览平台参考价，与创建任务同口径（起步价+里程+重量阶梯+夜间附加）")
    @PostMapping("/price/estimate")
    public Result<PriceDetailVO> estimatePrice(@RequestBody PriceEstimateDto dto) {
        return Result.success("预估成功", taskService.estimatePrice(dto));
    }

    @OperationLog("查询发布配置")
    @Operation(summary = "发布页配置", description = "重量计费类型、协商价下限等发布页展示配置")
    @GetMapping("/publish/config")
    public Result<PublishConfigVO> publishConfig() {
        PublishConfigVO vo = new PublishConfigVO();
        vo.setWeightTypes(billConfigService.getJson("WEIGHT_TYPES",
                new TypeReference<List<String>>() {}, List.of("TRANSPORT")));
        vo.setMinNegotiatedRate(billConfigService.getBigDecimal("MIN_NEGOTIATED_RATE", new BigDecimal("0.5")));
        return Result.success("获取成功", vo);
    }

    @OperationLog("查询任务列表")
    @Operation(summary = "获取当前用户的任务列表")
    @GetMapping("/list")
    public Result<TaskPageVO> listTasks(@RequestParam(defaultValue = "0") int page,
                                         @RequestParam(defaultValue = "20") int size) {
        Long userId = UserContext.getUserId();
        Page<Task> taskPage = taskService.getTasksByUser(userId, page, size);

        List<TaskVo> tasks = taskPage.getContent().stream()
                .map(this::toTaskVo)
                .toList();

        TaskPageVO vo = new TaskPageVO();
        vo.setTasks(tasks);
        vo.setCurrentPage(taskPage.getNumber());
        vo.setTotalPages(taskPage.getTotalPages());
        vo.setTotalElements(taskPage.getTotalElements());
        return Result.success("获取成功", vo);
    }

    @OperationLog("查询任务详情")
    @Operation(summary = "获取任务详情", description = "根据任务编号获取详细信息，包含航点列表",
            parameters = {@Parameter(name = "taskNum", description = "任务编号", required = true)})
    @GetMapping("/detail")
    public Result<TaskVo> getTaskDetail(@RequestParam String taskNum) {
        Long userId = UserContext.getUserId();
        Task task = taskService.getTaskByTaskNum(taskNum, userId);
        return Result.success("获取成功", toTaskVo(task));
    }

    private TaskVo toTaskVo(Task task) {
        TaskAssignment assignment = taskAssignmentRepository.findByTaskId(task.getId()).orElse(null);
        TaskVo vo = TaskVo.from(task, assignment);
        if (assignment != null) {
            userRepository.findById(assignment.getRiderId())
                .ifPresent(user -> vo.setRiderName(user.getUserName()));
        }
        MissionOrder order = orderRepository.findByTaskId(task.getId()).orElse(null);
        if (order != null) {
            vo.setOrderNum(order.getOrderNum());
            vo.setTotalAmount(order.getTotalAmount());
            vo.setTotalDistance(order.getTotalDistance());
            vo.setOrderStatus(order.getOrderStatus().name());
        }
        // 1B-9a 状态矩阵：任务状态×订单状态 → 操作提示
        vo.setActionHint(TaskActionHints.hint(task.getTaskStatus(),
                order != null ? order.getOrderStatus() : null));
        // 1B-4b：任务→设备映射（deviceId/liveState），用户端据此点亮「观看直播」入口
        var liveDevice = liveDeviceResolver.resolveForTask(task.getId());
        vo.setDeviceId(liveDevice.deviceId());
        vo.setLiveState(liveDevice.liveState());
        return vo;
    }

    @OperationLog("删除任务")
    @RateLimiter(limit = 5, windowSeconds = 60)
    @Operation(summary = "删除任务", description = "删除指定ID的任务，只能删除自己创建的任务",
            parameters = {@Parameter(name = "id", description = "任务数据库ID", required = true)})
    @DeleteMapping("/delete")
    public Result<Void> deleteTask(@RequestParam Long id) {
        Long userId = UserContext.getUserId();
        taskService.deleteTask(id, userId);
        return Result.success("任务删除成功");
    }

    @OperationLog("确认收货")
    @Operation(summary = "确认收货", description = "骑手完成任务后用户确认收货，订单完结",
            parameters = {@Parameter(name = "taskNum", description = "任务编号", required = true)})
    @PostMapping("/confirm")
    public Result<Void> confirmTask(@RequestParam String taskNum) {
        Long userId = UserContext.getUserId();
        taskService.userConfirmTask(taskNum, userId);
        return Result.success("确认收货成功");
    }
}