# drone-backend/src/main/java/com/uav/uav/controller/AppUavController.java
package com.uav.uav.controller;

import com.uav.uav.pojo.dto.UavDto;
import com.uav.server.result.Result;
import com.uav.uav.pojo.vo.UavVo;
import com.uav.server.annotation.OperationLog;
import com.uav.uav.service.AppUavService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

@Tag(name = "AppUav API")
@RestController
@RequestMapping("/appUav")
@Slf4j
public class AppUavController {

    @Autowired
    private AppUavService appUavService;

    // P0-6：无人机注册必须登录（移除调试残留的 @SkipJwt，匿名可抢注任意 djiId）
    @OperationLog("新增无人机")
    @Operation(
            summary = "新增无人机",
            description = "添加新的无人机信息（需登录）",
            responses = {
                    @ApiResponse(responseCode = "200", description = "新增成功"),
                    @ApiResponse(responseCode = "400", description = "参数错误")
            }
    )
    @PostMapping("/add")
    public Result<UavVo> addUav(@RequestBody UavDto uavDto) {
        UavVo uavVo = appUavService.tryToAddUav(uavDto);
        return Result.success(uavVo);
    }

    @OperationLog("更新无人机")
    @Operation(
            summary = "更新无人机信息",
            description = "更新无人机的基本信息，包括名称、在线状态等",
            responses = {
                    @ApiResponse(responseCode = "200", description = "更新成功"),
                    @ApiResponse(responseCode = "404", description = "无人机不存在")
            }
    )
    @PutMapping("/update")
    public Result<Void> updateOnlineStatus(@RequestBody UavDto uavDto) {
        appUavService.updateUav(uavDto);
        return Result.success("更新成功");
    }
}