# spec/openapi/drone-backend.openapi.json
{
  "components" : {
    "schemas" : {
...
      "AdminPageVoAdminTaskVo" : {
        "description" : "管理端分页信封（page 从 0 开始）",
        "properties" : {
          "content" : {
            "description" : "当前页数据列表",
            "items" : {
              "$ref" : "#/components/schemas/AdminTaskVo"
            },
            "type" : "array"
          },
          "page" : {
            "description" : "当前页码（从 0 开始）",
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "description" : "每页条数",
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "description" : "总记录数",
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "description" : "总页数（由总记录数与每页条数计算）",
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "AdminTaskVo" : {
        "description" : "管理端任务信息（含任务状态与关联订单状态）",
        "properties" : {
          "actionHint" : {
            "description" : "操作提示文案（由任务状态 × 订单状态计算，客户端可直接渲染）",
            "type" : "string"
          },
          "completeNote" : {
            "description" : "飞手完成说明（未完成或飞手未填写时为 null）",
            "type" : "string"
          },
          "createTime" : {
            "description" : "创建时间，格式 yyyy-MM-dd HH:mm:ss",
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "description" : "任务描述",
            "type" : "string"
          },
          "id" : {
            "description" : "任务ID",
            "format" : "int64",
            "type" : "integer"
          },
          "orderNum" : {
            "description" : "关联订单号（无关联订单时为 null）",
            "type" : "string"
          },
          "orderStatus" : {
            "description" : "关联订单状态枚举名（PENDING/PAID/CANCELLED/REFUNDED/COMPLETED/WAITING_CONFIRM/DISPUTED；无关联订单时为 null）",
            "type" : "string"
          },
          "orderStatusCode" : {
            "description" : "关联订单状态码（0 待支付 / 1 已支付 / 2 已取消 / 3 已退款 / 4 已完成 / 5 待确认完成 / 6 争议中；无关联订单时为 null）",
            "format" : "int32",
            "type" : "integer"
          },
          "orderStatusDesc" : {
            "description" : "关联订单状态中文描述（无关联订单时为 null）",
            "type" : "string"
          },
          "ownerName" : {
            "description" : "发布用户名称",
            "type" : "string"
          },
          "reward" : {
            "description" : "任务奖励金额，单位：元（服务端按航点距离计价，客户端传入的 reward 仅作参考）",
            "format" : "double",
            "type" : "number"
          },
          "riderName" : {
            "description" : "接单飞手名称（未接单时为 null）",
            "type" : "string"
          },
          "taskName" : {
            "description" : "任务名称",
            "type" : "string"
          },
          "taskNum" : {
            "description" : "任务编号",
            "type" : "string"
          },
          "taskStatus" : {
            "description" : "任务状态（IDLE 空闲中 / IN_PROGRESS 执行中 / COMPLETED 执行完毕）",
            "type" : "string"
          },
          "taskStatusDesc" : {
            "description" : "任务状态中文描述",
            "type" : "string"
          },
          "taskTime" : {
            "description" : "任务期望执行时间，格式 yyyy-MM-dd HH:mm:ss",
            "format" : "date-time",
            "type" : "string"
          },
          "totalAmount" : {
            "description" : "关联订单总金额，单位：元（无关联订单时为 null）",
            "type" : "number"
          },
          "totalDistance" : {
            "description" : "关联订单总里程，单位：米（无关联订单时为 null）",
            "type" : "number"
          },
          "updateTime" : {
            "description" : "更新时间，格式 yyyy-MM-dd HH:mm:ss",
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "description" : "发布用户ID",
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PriceDetailVO" : {
        "properties" : {
          "distanceKm" : {
            "type" : "number"
          },
          "isNight" : {
            "type" : "boolean"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/PriceItem"
            },
            "type" : "array"
          },
          "needManualQuote" : {
            "type" : "boolean"
          },
          "total" : {
            "type" : "number"
          },
          "weightKg" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "PriceEstimateDto" : {
        "properties" : {
          "taskTime" : {
            "type" : "string"
          },
          "taskType" : {
            "enum" : [ "AERIAL_PHOTO", "MONITORING", "SURVEY", "TRANSPORT" ],
            "type" : "string"
          },
          "waypoints" : {
            "items" : {
              "$ref" : "#/components/schemas/WaypointDto"
            },
            "type" : "array"
          },
          "weight" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "PriceItem" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ResultAdminPageVoAdminTaskVo" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/AdminPageVoAdminTaskVo",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultPriceDetailVO" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/PriceDetailVO",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultTaskPageVO" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/TaskPageVO",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ResultTaskVo" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/TaskVo",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultUavVo" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/UavVo",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultVoid" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "TaskDto" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "reward" : {
            "format" : "double",
            "type" : "number"
          },
          "taskName" : {
            "type" : "string"
          },
          "taskTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "type" : {
            "enum" : [ "AERIAL_PHOTO", "MONITORING", "SURVEY", "TRANSPORT" ],
            "type" : "string"
          },
          "waypoints" : {
            "items" : {
              "$ref" : "#/components/schemas/WaypointDto"
            },
            "type" : "array"
          },
          "weight" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "TaskPageVO" : {
        "properties" : {
          "currentPage" : {
            "format" : "int32",
            "type" : "integer"
          },
          "tasks" : {
            "items" : {
              "$ref" : "#/components/schemas/TaskVo"
            },
            "type" : "array"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "TaskVo" : {
        "properties" : {
          "acceptTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "actionHint" : {
            "type" : "string"
          },
          "completeNote" : {
            "type" : "string"
          },
          "createTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "deviceId" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "liveState" : {
            "type" : "string"
          },
          "needManualQuote" : {
            "type" : "boolean"
          },
          "orderNum" : {
            "type" : "string"
          },
          "orderStatus" : {
            "type" : "string"
          },
          "priceDetail" : {
            "type" : "string"
          },
          "referencePrice" : {
            "type" : "number"
          },
          "riderId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "riderName" : {
            "type" : "string"
          },
          "taskName" : {
            "type" : "string"
          },
          "taskNum" : {
            "type" : "string"
          },
          "taskStatus" : {
            "enum" : [ "COMPLETED", "IDLE", "IN_PROGRESS" ],
            "type" : "string"
          },
          "taskTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "taskType" : {
            "enum" : [ "AERIAL_PHOTO", "MONITORING", "SURVEY", "TRANSPORT" ],
            "type" : "string"
          },
          "totalAmount" : {
            "type" : "number"
          },
          "totalDistance" : {
            "type" : "number"
          },
          "updateTime" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "waypoints" : {
            "items" : {
              "$ref" : "#/components/schemas/WaypointVo"
            },
            "type" : "array"
          },
          "weight" : {
            "format" : "double",
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
      "UavDto" : {
        "description" : "Parameters for Add Request",
        "properties" : {
          "controllerModel" : {
            "description" : "controller_model",
            "type" : "string"
          },
          "djiId" : {
            "description" : "dji_id",
            "type" : "string"
          },
          "onlineStatus" : {
            "description" : "online_status",
            "type" : "string"
          },
          "uavName" : {
            "description" : "uav_name",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UavVo" : {
        "properties" : {
          "controllerModel" : {
            "description" : "控制器型号",
            "type" : "string"
          },
          "djiId" : {
            "description" : "设备ID(DJI ID)",
            "type" : "string"
          },
          "id" : {
            "description" : "id",
            "format" : "int64",
            "type" : "integer"
          },
          "isAvailable" : {
            "description" : "是否可用(1可用,0禁用)",
            "type" : "string"
          },
          "uavName" : {
            "description" : "uavName",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "WaypointDto" : {
        "properties" : {
          "altitude" : {
            "format" : "double",
            "type" : "number"
          },
          "latitude" : {
            "format" : "double",
            "type" : "number"
          },
          "longitude" : {
            "format" : "double",
            "type" : "number"
          },
          "orderIndex" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "WaypointVo" : {
        "properties" : {
          "altitude" : {
            "format" : "double",
            "type" : "number"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "latitude" : {
            "format" : "double",
            "type" : "number"
          },
          "longitude" : {
            "format" : "double",
            "type" : "number"
          },
          "orderIndex" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "Bearer Authentication" : {
        "bearerFormat" : "JWT",
        "description" : "请输入JWT Token，格式：直接输入token字符串即可",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "无人机项目web后端接口文档",
    "title" : "无人机项目接口文档",
    "version" : "1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/admin/tasks" : {
      "get" : {
        "description" : "支持按任务状态（枚举名：IDLE/IN_PROGRESS/COMPLETED）与任务号精确过滤；含订单双状态与操作提示",
        "operationId" : "listTasks_1",
        "parameters" : [ {
          "description" : "任务状态过滤：IDLE / IN_PROGRESS / COMPLETED",
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "任务编号精确过滤",
          "in" : "query",
          "name" : "taskNum",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "每页条数（默认 20，上限 100）",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "description" : "页码（从 0 起）",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultAdminPageVoAdminTaskVo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "全平台任务分页列表",
        "tags" : [ "Admin Query API" ]
      }
    },
...
    "/appUav/add" : {
      "post" : {
        "description" : "添加新的无人机信息（需登录）",
        "operationId" : "addUav",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UavDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultUavVo"
                }
              }
            },
            "description" : "新增成功"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultUavVo"
                }
              }
            },
            "description" : "参数错误"
          }
        },
        "summary" : "新增无人机",
        "tags" : [ "AppUav API" ]
      }
    },
    "/appUav/update" : {
      "put" : {
        "description" : "更新无人机的基本信息，包括名称、在线状态等",
        "operationId" : "updateOnlineStatus",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UavDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultVoid"
                }
              }
            },
            "description" : "更新成功"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultVoid"
                }
              }
            },
            "description" : "无人机不存在"
          }
        },
        "summary" : "更新无人机信息",
        "tags" : [ "AppUav API" ]
      }
    },
...
    "/rider/task/detail" : {
      "get" : {
        "description" : "飞手查看任意任务的详细信息（无需任务归属）",
        "operationId" : "getTaskDetail_1",
        "parameters" : [ {
          "description" : "任务编号",
          "in" : "query",
          "name" : "taskNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTaskVo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "任务详情",
        "tags" : [ "Rider API" ]
      }
    },
    "/task/confirm" : {
      "post" : {
        "description" : "骑手完成任务后用户确认收货，订单完结",
        "operationId" : "confirmTask",
        "parameters" : [ {
          "description" : "任务编号",
          "in" : "query",
          "name" : "taskNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "确认收货",
        "tags" : [ "Task API" ]
      }
    },
    "/task/create" : {
      "post" : {
        "description" : "创建新任务，包含任务类型、航点信息等",
        "operationId" : "createTask",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TaskDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTaskVo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "创建任务",
        "tags" : [ "Task API" ]
      }
    },
    "/task/delete" : {
      "delete" : {
        "description" : "删除指定ID的任务，只能删除自己创建的任务",
        "operationId" : "deleteTask",
        "parameters" : [ {
          "description" : "任务数据库ID",
          "in" : "query",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "删除任务",
        "tags" : [ "Task API" ]
      }
    },
    "/task/detail" : {
      "get" : {
        "description" : "根据任务编号获取详细信息，包含航点列表",
        "operationId" : "getTaskDetail",
        "parameters" : [ {
          "description" : "任务编号",
          "in" : "query",
          "name" : "taskNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTaskVo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "获取任务详情",
        "tags" : [ "Task API" ]
      }
    },
...
    "/task/list" : {
      "get" : {
        "operationId" : "listTasks",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTaskPageVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "获取当前用户的任务列表",
        "tags" : [ "Task API" ]
      }
    },
    "/task/price/estimate" : {
      "post" : {
        "description" : "发布前预览平台参考价，与创建任务同口径（起步价+里程+重量阶梯+夜间附加）",
        "operationId" : "estimatePrice",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PriceEstimateDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultPriceDetailVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "参考价预估",
        "tags" : [ "Task API" ]
      }
    },
...
  },
  "security" : [ {
    "Bearer Authentication" : [ ]
  } ],
  "servers" : [ {
    "description" : "相对基址：由部署环境（网关/反向代理）决定",
    "url" : "/"
  } ],
...
}